pyproject.toml
[project]
name = "netclab-examples"
version = "0.0.1"
authors = [
    { name = "mbakalarski at github" }
]
description = "Network tests on netclab"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: POSIX :: Linux",
]
license = "MIT"
license-files = ["LICENSE"]

dependencies = [
    "pytest>=8.3.4",
    "snappi>=1.19.1",
    "kubernetes>=31.0.0",
    "python-dotenv>=1.0.1",
    "pyats>=24.11",
]


[project.urls]
Homepage = "https://github.com/mbakalarski/netclab-examples"


[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/netclab-examples"]
